* releases: (46 commits)
  REL: 0.14 release: update whats_new and version
  MAINT Update mailmap
  DOC: fix CSS bug
  DOC: update documentation for release
  MISC: deprection is in 2 releases
  MAINT: randn on float is deprecated
  TST: avoid nose running sklearn.test as a test
  MISC: fix wrong timing in example
  FIX integer types in Ward clustering
  TST: avoid a crash in Windows + Anaconda Py3.3
  FIX search and replace misstake
  DOC put the narrative documentation of roc_curve and roc_auc_score in one place
  ENH more explicit name for auc + consistency for scorer, fix #2096
  DOC: button layout tweak
  Website: bottom buttons
  DOC: bigger menu fonts
  DOC: layout tweaks
  Polishing on "Who's using scikit-learn"
  FIX typo in testimonials
  ENH testimonials img are now centered.
  ...

+++ sklearn/tests/test_random_projection.py
@@ -53,7 +53,7 @@ def densify(matrix):
 
 
 n_samples, n_features = (10, 1000)
-n_nonzeros = n_samples * n_features / 100.
+n_nonzeros = int(n_samples * n_features / 100.)
 data, data_csr = make_sparse_random_data(n_samples, n_features, n_nonzeros)
